Converts SPPS-syntax either from a SPSS-syntax file .sps or read from a SPSS-output-file .spv (using spv2sps) into jamovi / jmv-syntax

The function expects that either the parameter vecSPS is used (with the output from spv2sps) or that both fleSPS and fleSAV are given.

sps2jmv(vecSPS = c(), fleSPS = "", fleSAV = "", runMsV = FALSE)

Arguments

vecSPS

vector with SPSS-commands, the format is supposed to be the same as what is output from spv2sps: a character vector containing one command per element, and the attribute "datafile" that contains a string pointing to the location of the SPSS-data-file (incl. path; if not found at the path or if no path is given, the file is searched in the current directory)

fleSPS

containing the location of a SPSS-syntax-file (.sps; incl. path)

fleSAV

containing the location of a SPSS-data-file (.sav; incl. path)

runMsV

run analyses even though there are variables missing; if the value is set to FALSE (default), the whole analysis is preceded by "#" which would cause it not to be evaluated with "eval(parse(text = ...))"; if the value is set to TRUE, only the variable that is not found in the data set is excluded and the analyses are conducted with the remaining variables (which will lead to results that are different from those originally conducted in SPSS)

Value

list with jamovi / jmv-analysis-function-calls and the data set stored in the attribute "dataset"

Examples

if (FALSE) {
library(jmvReadWrite);
}