Twenty-five personality self-report items taken from the International Personality Item Pool

The data set contains responses from 250 participants filling in twenty-five personality self-report items taken from the International Personality Item Pool (https://ipip.ori.org) as part of the Synthetic Aperture Personality Assessment (SAPA) web-based personality assessment (https://sapa-project.org) project. The 25 items are organized by five putative factors: Agreeableness (A1 to A5), Conscientiousness (C1 to C5), Extraversion (E1 to E5), Neuroticism (N1 to N5), and Openness (N1 to N5). The items were short phrases that the respondent should answer by indicating how accurately the statement describes their typical behaviour or attitude. Responses were collected using a 6-point scale: 1 - Very inaccurate, 2 - Moderately inaccurate, 3 - Slightly inaccurate, 4 - Slightly accurate, 5 - Moderately accurate, 6 - Very accurate.

bfi_sample

Format

A data.frame with 254 rows (250 original respondents, 4 manually generated for testing) and 33 variables

ID

character

Respondent ID
A1

integer

Am indifferent to the feelings of others. (reversed)
A2

integer

Inquire about others' well-being.
A3

integer

Know how to comfort others.
A4

integer

Love children.
A5

integer

Make people feel at ease.
C1

integer

Am exacting in my work.
C2

integer

Continue until everything is perfect.
C3

integer

Do things according to a plan.
C4

integer

Do things in a half-way manner. (reversed)
C5

integer

Waste my time. (reversed)
E1

integer

Don't talk a lot. (reversed)
E2

integer

Find it difficult to approach others. (reversed)
E3

integer

Know how to captivate people.
E4

integer

Make friends easily.
E5

integer

Take charge.
N1

integer

Get angry easily.
N2

integer

Get irritated easily.
N3

integer

Have frequent mood swings.
N4

integer

Often feel blue.
N5

integer

Panic easily.
O1

integer

Am full of ideas.
O2

integer

Avoid difficult reading material. (reversed)
O3

integer

Carry the conversation to a higher level.
O4

integer

Spend time reflecting on things.
O5

integer

Will not probe deeply into a subject. (reversed)
gender

factor

Gender of the respondent (female, male)
age

integer

Age of the respondent (years)
AD

numeric

Exponent of age (computed: EXP(age))
AF

factor

Random data (for testing)
AG

factor

Random data (for testing)
age_tr

factor

Age of the respondent (transformed, as decades: 1 - 10-19, 2 - 20-29, 3 - 30-39, 4 - 40-49, 5 - 50-59, 6 - 60 and over)
ID2

character

Respondent ID (for testing; "A" + random-generated 5-digit-code)